chore: remove brotli compression and pkg dependency
This commit completes the removal of the pkg-based compression system
and simplifies the CLI loader to directly require the built bundle.

Changes:
- Remove compress-cli.mjs scripts (no longer needed)
- Simplify src/index.mts to directly load cli.js (no decompression)
- Add copy-cli.mjs scripts to copy built CLI to dist/
- Update build scripts to use new copy approach
- Update esbuild configs to output uncompressed cli.js
- Update verify-package scripts for new structure

The CLI now ships as an uncompressed bundle, removing the ~10MB→1.7MB
compression at the cost of larger npm package size. This simplifies
the build process and removes the pkg dependency.

Related to previous commit: 3b99270a1 (chore: remove pkg dependency and related build scripts)
