+++ packages/cli-with-sentry/scripts/verify-package.mjs
@@ -85,7 +85,7 @@ async function validate() {
   }
 
   // Check dist files exist and validate Sentry integration.
-  const distFiles = ['index.js', 'cli.js.bz', 'shadow-npm-inject.js']
+  const distFiles = ['index.js', 'cli.js', 'shadow-npm-inject.js']
   for (const file of distFiles) {
     logger.info(`Checking dist/${file}...`)
     const filePath = path.join(packageRoot, 'dist', file)
