+++ packages/cli/scripts/verify-package.mjs
@@ -77,7 +77,7 @@ async function validate() {
   }
 
   // Check dist files exist.
-  const distFiles = ['index.js', 'cli.js.bz', 'shadow-npm-inject.js']
+  const distFiles = ['index.js', 'cli.js', 'shadow-npm-inject.js']
   for (const file of distFiles) {
     logger.info(`Checking dist/${file}...`)
     const filePath = path.join(packageRoot, 'dist', file)
